+++ metadata.rb
@@ -3,7 +3,7 @@ maintainer        "Opscode, Inc."
 maintainer_email  "cookbooks@opscode.com"
 license           "Apache 2.0"
 description       "Installs mercurial"
-version           "0.8.1"
+version           "1.1.4"
 
 recipe "mercurial", "Installs mercurial"
 
@@ -12,3 +12,4 @@ recipe "mercurial", "Installs mercurial"
 end
 
 depends           "windows"
+depends           "python"