+++ providers/default.rb
@@ -1,22 +1,22 @@
 action :sync do
-  ::File.delete(hgupFile) if ::File.exist?(hgupFile)
-  execute "sync repository #{new_resource.path}" do    
+  ::File.delete(hgup_file) if ::File.exist?(hgup_file)
+  execute "sync repository #{new_resource.path}" do
     not_if "hg identify #{new_resource.path}"
-    command "hg clone --rev #{new_resource.reference} #{hgConnectionCommand} #{new_resource.repository} #{new_resource.path} && touch #{hgupFile}"
-    creates hgupFile
+    command "hg clone --rev #{new_resource.reference} #{hg_connection_command} #{new_resource.repository} #{new_resource.path} && touch #{hgup_file}"
+    creates hgup_file
     notifies :run, "execute[set ownership]"
     notifies :run, "execute[set permissions]"
   end
   execute "check incoming changes" do
-    command "hg incoming --rev #{new_resource.reference} #{hgConnectionCommand}  #{new_resource.repository} && touch #{hgupFile} || true"
+    command "hg incoming --rev #{new_resource.reference} #{hg_connection_command}  #{new_resource.repository} && touch #{hgup_file} || true"
     cwd new_resource.path
-    creates hgupFile
+    creates hgup_file
     notifies :run, "execute[pull]"
   end
   execute "pull" do
-    command "hg pull --rev #{new_resource.reference} #{hgConnectionCommand} #{new_resource.repository}"
+    command "hg pull --rev #{new_resource.reference} #{hg_connection_command} #{new_resource.repository}"
     cwd new_resource.path
-    only_if { ::File.exist?(hgupFile) }
+    only_if { ::File.exist?(hgup_file) }
     action :nothing
     notifies :run, "execute[update]"
   end
@@ -31,24 +31,25 @@ action :sync do
     command "chown -R #{new_resource.owner}:#{new_resource.group} #{new_resource.path}"
     action :nothing
   end
-  execute "set permissions" do
-    command "chmod -R #{new_resource.mode} #{new_resource.path}"
-    action :nothing
+  if new_resource.mode then
+    execute "set permissions" do
+      command "chmod -R #{new_resource.mode} #{new_resource.path}"
+    end
   end
-  if ::File.exist?(hgupFile)
+  if ::File.exist?(hgup_file)
     new_resource.updated_by_last_action(true)
-    ::File.delete(hgupFile)
+    ::File.delete(hgup_file)
   else
     new_resource.updated_by_last_action(false)
   end
 end
- 
+
 action :clone do
-  ::File.delete(hgupFile) if ::File.exist?(hgupFile)
+  ::File.delete(hgup_file) if ::File.exist?(hgup_file)
   execute "clone repository #{new_resource.path}" do
-    command "hg clone --rev #{new_resource.reference} #{hgConnectionCommand} #{new_resource.repository} #{new_resource.path} && touch #{hgupFile}"
+    command "hg clone --rev #{new_resource.reference} #{hg_connection_command} #{new_resource.repository} #{new_resource.path} && touch #{hgup_file}"
     not_if "hg identify #{new_resource.path}"
-    creates hgupFile
+    creates hgup_file
     notifies :run, "execute[set permission]"
     notifies :run, "execute[set ownership]"
   end
@@ -56,13 +57,14 @@ action :clone do
     command "chown -R #{new_resource.owner}:#{new_resource.group} #{new_resource.path}"
     action :nothing
   end
-  execute "set permission" do
-    command "chmod -R #{new_resource.mode} #{new_resource.path}"
-    action :nothing
+  if new_resource.mode then
+    execute "set permission" do
+      command "chmod -R #{new_resource.mode} #{new_resource.path}"
+    end
   end
-  if ::File.exist?(hgupFile)
+  if ::File.exist?(hgup_file)
     new_resource.updated_by_last_action(true)
-    ::File.delete(hgupFile)
+    ::File.delete(hgup_file)
   else
     new_resource.updated_by_last_action(false)
   end