+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## v1.1.4:
+
+* [COOK-2278] - Install mercurial using pip
+* [COOK-2279] - typo in README.md
+
+## v1.1.2:
+
+* [COOK-2033] - only set recursive permissions if mode is used in
+  mercurial resource
+
+## v1.1.0:
+
+* [COOK-1945] - cleanup provider
+* [COOK-1946] - windows support
+
 ## v1.0.0:
 
 * [COOK-1373] - README example correction